♻️ refactor: upgrade tree-sitter to 0.26.5 and all language grammars
- Upgrade tree-sitter from 0.23 to 0.26.5
- Upgrade tree-sitter-rust to 0.24.0
- Upgrade tree-sitter-python to 0.25.0
- Upgrade tree-sitter-javascript to 0.25.0
- Upgrade tree-sitter-typescript to 0.23.2
- Upgrade tree-sitter-c to 0.24.1
- Upgrade tree-sitter-cpp to 0.23.4
- Upgrade tree-sitter-c-sharp to 0.23.1
- Upgrade tree-sitter-go to 0.25.0
- Upgrade tree-sitter-java to 0.23.5

Breaking change: tree-sitter 0.26 changed named_child() parameter
from usize to u32. Fixed all type mismatches in extract_docstring()
methods across Rust, JavaScript, C#, Go, Java extractors and
the extract_c_style_doc() helper function.

All 195 tests pass successfully.

+++ src/chunker/extractor.rs
@@ -853,10 +853,10 @@ impl LanguageExtractor for GoExtractor {
         // Go uses // comments before declarations
         let parent = node.parent()?;
         let node_index = (0..parent.named_child_count())
-            .find(|&i| parent.named_child(i).map(|c| c.id()) == Some(node.id()))?;
+            .find(|&i| parent.named_child(i as u32).map(|c| c.id()) == Some(node.id()))?;
 
         if node_index > 0 {
-            if let Some(prev) = parent.named_child(node_index - 1) {
+            if let Some(prev) = parent.named_child((node_index - 1) as u32) {
                 if prev.kind() == "comment" {
                     return prev.utf8_text(source).ok().map(String::from);
                 }
@@ -939,10 +939,10 @@ impl LanguageExtractor for JavaExtractor {
         // Java uses /** */ Javadoc comments
         let parent = node.parent()?;
         let node_index = (0..parent.named_child_count())
-            .find(|&i| parent.named_child(i).map(|c| c.id()) == Some(node.id()))?;
+            .find(|&i| parent.named_child(i as u32).map(|c| c.id()) == Some(node.id()))?;
 
         if node_index > 0 {
-            if let Some(prev) = parent.named_child(node_index - 1) {
+            if let Some(prev) = parent.named_child((node_index - 1) as u32) {
                 if prev.kind() == "block_comment" || prev.kind() == "comment" {
                     if let Ok(text) = prev.utf8_text(source) {
                         if text.trim_start().starts_with("/**") {
@@ -1005,10 +1005,10 @@ fn find_identifier(node: Node, source: &[u8]) -> Option<String> {
 fn extract_c_style_doc(node: Node, source: &[u8]) -> Option<String> {
     let parent = node.parent()?;
     let node_index = (0..parent.named_child_count())
-        .find(|&i| parent.named_child(i).map(|c| c.id()) == Some(node.id()))?;
+        .find(|&i| parent.named_child(i as u32).map(|c| c.id()) == Some(node.id()))?;
 
     if node_index > 0 {
-        if let Some(prev) = parent.named_child(node_index - 1) {
+        if let Some(prev) = parent.named_child((node_index - 1) as u32) {
             if prev.kind() == "comment" || prev.kind() == "block_comment" {
                 if let Ok(text) = prev.utf8_text(source) {
                     let trimmed = text.trim_start();
